+++ src/Provider/Psr15/RelayServiceProvider.php
@@ -46,6 +46,7 @@ class RelayServiceProvider implements ServiceProviderInterface
         $app['handler.queue'] = [];
         
         $app['uSilex.httpHandler'] = function($app) {
+
             $resolver = function($entry) use ($app) {
                 return is_string($entry) ? $app[$entry] : $entry;
             };