+++ CHANGELOG.md
@@ -5,6 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.1.0]
+
+### Added
+
+- ContainerAware trait
+- PSR-11 implementation in Application
+
+
+### Fixed
+
+- Removed Application Constructor
+- Remove dependency from ['debug'] in Psr7 DiactorosServiceProvider
+
 ## [1.0.1] 2018-10-02
 
 ### Added
@@ -23,5 +36,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 First release. 
 
 
-[Unreleased]: https://github.com/linkeddatacenter/uSilex/compare/1.0.1...HEAD
+[Unreleased]: https://github.com/linkeddatacenter/uSilex/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/linkeddatacenter/uSilex/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/linkeddatacenter/uSilex/compare/1.0.0...1.0.1